requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1 